Further work on parameters, can now actually list all values and ranges

+++ Source/LrControlApi/Common/Range.cs
@@ -2,13 +2,13 @@
 {
     public class Range
     {
-        public Range(int minimum, int maximum)
+        public Range(double minimum, double maximum)
         {
             Minimum = minimum;
             Maximum = maximum;
         }
 
-        public int Minimum { get; }
-        public int Maximum { get; }
+        public double Minimum { get; }
+        public double Maximum { get; }
     }
 }
